+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 * feat(Calendar): add `weekNumbers` prop
 * fix(Editor): support all heading levels by default
 * docs(i18n): improve locale documentation for date/time formatting
+* feat(Timeline): add `select` event
 
 ## [2.2.1](https://github.com/bitrix24/b24ui/compare/v2.1.17...v2.2.1) (2025-12-18)